% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascAnon.R
\name{ascAnon}
\alias{ascAnon}
\title{Remove potentially identifying information from SRR *ASC file.}
\usage{
ascAnon(file, newName = NULL)
}
\arguments{
\item{file}{Path to SRR *ASC file.}

\item{newName}{Name to use for newly anonymized *ASC file. If an extension is included it will
be ignored and replaced with "asc". If NULL (the default) will use the same name as the
source *ASC file, but with "-anon" appended. In either case the same base filename will be
used as the name of the *EDF filename in the "CONVERTED FROM" line of the anonynmized *ASC
file.}
}
\value{
Filename and path to anonymized file.
}
\description{
Remove potentially identifying information from header of an SRR *ASC file. Write
    the new file to the current working directory. This is a superficial
    method of de-identification or anonymization, but it should do for most cases.
}
\details{
By default, newly created anonymous *ASC file will will use the same name as the source
    *ASC file, but with "-anon" appended. In this case, the same base filename will be used as
    the name of the *EDF filename in the "CONVERTED FROM" line of the anonynmized *ASC
    file. This is problematic in that, under our usual operating procedures, the filename itself
    usually consists of an identifier that is uniquely traceable to the participant. In order to
    make the file truly anonymous, the newName argument must be used.

    If newName is provided, that will be used as the basename for the EDF file in the CONVERTED
    FROM line, as well as the output file name.
}
\examples{
file <- system.file("extdata/1950168.asc", package="FDBeye")
ascAnon(file)
ascAnon(file, newName="fred")
}
\author{
Dave Braze <davebraze@gmail.com>
}

