% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_y.R
\name{cat_lines1}
\alias{cat_lines1}
\title{Categorize fixations into appropriate text lines}
\usage{
cat_lines1(params, fit_it = TRUE, data, start_pts, k_bounds, o_bounds,
  s_bounds, den_sd_cutoff, den_ratio_cutoff, num_checkFirst = 5,
  num_checkLast = 10)
}
\arguments{
\item{params}{parameters (slope, offset, sd) for optimization}

\item{fit_it}{TRUE -> return fit measure, FALSE -> return fit information}

\item{data}{fixation data frame containing at least x_pos, y_pos}

\item{start_pts}{dataframe contatining starting point of each base line (x_pos, y_pos, trial_num)}

\item{k_bounds}{boundary of slope (default: [-0.1, 0.1])}

\item{o_bounds}{boundary of offset (default: [-0.5*dist of adjacent text lines, 0.5*dist of
adjacent text line])}

\item{s_bounds}{boundary of sd (default: [1, 20])}

\item{den_sd_cutoff}{cutoff threshold for density; If den_sd_cutoff is Inf, use
mean(inv_dnorm(exp(data_den_max))) + 3*sd(inv_dnorm(exp(data_den_max))) as cutoff (99.7\% are
accepted)}

\item{den_ratio_cutoff}{cutoff threshold for density ratio (ratio between the maximum density and
second maximum density)}

\item{num_checkFirst}{number of starting fixations used for checking before reading bound
(default 5)}

\item{num_checkLast}{number of ending fixations used for checking end of reading bound (default
10)}
}
\value{
new data frame including fixations, fit measures, and fitted lines information
}
\description{
This function follows closely upon that described in Cohen's paper
}
\details{
This function optimizes the slope, offset and sd for all lines of fixations. It uses -sum(data_den_max)
 as fit measure for optimization
}
\author{
Tao Gong <gtojty@gmail.com>
}

