% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_y.R
\name{cat_lines2}
\alias{cat_lines2}
\title{Categorize fixations into appropriate text lines}
\usage{
cat_lines2(params, fit_it = TRUE, data, start_pts, k_bounds, o_bounds,
  s_bounds, den_sd_cutoff, den_ratio_cutoff, num_checkFirst = 5,
  num_checkLast = 10)
}
\arguments{
\item{params}{parameters (slope, vert_offset, sd) for optimization}

\item{fit_it}{TRUE -> return fit measure, FALSE -> return fit information}

\item{data}{fixation data frame containing at least x_pos, y_pos}

\item{start_pts}{starting point of each base line (x_pos, y_pos)}

\item{k_bounds}{boundary of slope}

\item{o_bounds}{boundary of offset}

\item{s_bounds}{boundary of sd}

\item{den_sd_cutoff}{cutoff threshold for density}

\item{den_ratio_cutoff}{cutoff threshold for density ratio (ratio between the maximum density and
second maximum density)}

\item{num_checkFirst}{number of starting fixations used for checking before reading bound
(default 5)}

\item{num_checkLast}{number of ending fixations used for checking end of reading bound (default
10)}
}
\value{
new data frame including fixations, fit measures, and fitted lines information
}
\description{
Function to optimize slope, offset and sd for all lines of fixations (extended from
    original paper)This function uses one slope, ver_offset, sd to fit all lines, and use sum(min
    y_diff) as target measure for optimization uses.
}
\details{
TAO, PLEASE PROVIDE DETAILS
}
\author{
Tao Gong <gtojty@gmail.com>
}

