% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EToralreading}
\alias{EToralreading}
\title{Data: Eye movements (fixations) over text.}
\format{A data.frame with 6676 rows and 8 variables:
\describe{
    \item{subj}{Subject identifier.}
    \item{eye}{The eye being tracked. Always "RIGHT".}
    \item{xpos}{Screen coordinates of the fixation; its (average) x position in pixels.}
    \item{ypos}{Screen coordinates of the fixation; its (average) y position in pixels. Note that the origin is in the upper left hand corner.}
    \item{stime}{Start time of fixation, in milliseconds from the beginning of the trial.}
    \item{etime}{End time of fixation, in milliseconds from the beginning of the trial.}
    \item{label}{Hand assigned labels corresponding to the nominal line of text for each fixation (line 1, 2, 3, etc).
                 Assignment was done by first identifying the "retrace" (return from the end of line n to the beginning
                 of line n+1) after each line of text. Then, fixations
                 before the first retrace were assigned to line 1; those between the first and second retrace were
                 assigned to line 2, etc. Labels with values like "x.5" indicate fixations that occur within a retrace.}
    \item{story}{Identifier for the text being read (1,2, or 3). Bit map images for each text can be found as
                    system.file("extdata/story01.png", package="FDBeye"), for example. Coordinate values for fixations
                    correspond to pixel locations in the bitmaps.}
}}
\source{
Data collected as part of U.S. National Institutes of Health grant R01 HD071988 to
    Haskins Laboratories.
}
\usage{
EToralreading
}
\description{
Nine different readers reading the same three stories. Some included data is
    poor quality.
}
\details{
Data were collected using an SR Research Eyelink 1000+ eye tracker with remote camera
    configuration and head stabilization. Data were extracted from EDF files as
    "fixation reports" using the SRR Data Viewer software.
}
\keyword{datasets}

