% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixPlot.R
\name{fixPlot}
\alias{fixPlot}
\title{Plot Gaze Fixation Data}
\usage{
fixPlot(data, bgimage, bgalpha = 0.33, xyMap = ggplot2::aes_string(x =
  "xpos", y = "ypos"), pointMap = ggplot2::aes_string(size = "dur"),
  pointAlpha = 0.5, mar = c(0, 0, 0, 0), showRug = FALSE, rugMap = NULL,
  rugColor = "black", rugAlpha = 0.33, rugSize = 2, ...,
  showPlot = TRUE)
}
\arguments{
\item{data}{A data.frame containing fixations. Must include x and y positions with column names
"x" and "y".}

\item{bgimage}{Path to background image. Must be a PNG file. Usually this will correspond to the visual stimulus
being viewed by subjects while their eye movements were recorded.}

\item{bgalpha}{Alpha level for background layer. Currently not working.}

\item{xyMap}{Aesthetic mapping (ggplot2::aes_string()) for x and y coordinates. Passed to ggplot().}

\item{pointMap}{Additional aesthetics specific to points. Passed to geom_point().}

\item{pointAlpha}{Set point transparency.}

\item{mar}{A 4 vector for margin adjustment: top, right, bottom, left. Use positive values move
margins toward center of plot (trim the display area). Use negative values to expand the
display area beyond the bitmap.}

\item{showRug}{Logical indicating whether to include rug plots for x and y margins. Defaults is
FALSE.}

\item{rugMap}{Additional aesthetics specific to rugs. Passed to geom_rug().}

\item{rugColor}{Set rug color.}

\item{rugAlpha}{Set rug transparency.}

\item{rugSize}{Set rug size.}

\item{...}{Additional arguments passed to ggplot2::geom_point().}

\item{showPlot}{Logical indicating whether to display the plot. Defaults to TRUE.}
}
\value{
A ggplot2 object.
}
\description{
Plot gaze fixation data overlayed on a bitmap image. This is a convenience wrapper
    around ggplot2::ggplot().
}
\details{
This is a convenience function for using ggplot() to lay fixation data over a
    background bitmap. At present, only PNG files are supported for background images. Usually,
    these will correspond to the visual stimulus viewed by subjects while gaze was
    recorded. Function returns a ggplot2 object which can be further modified by ggplot2 geoms.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

