% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeReport.R
\name{fixReport}
\alias{fixReport}
\title{Make a fixation report from an ELascii object.}
\usage{
fixReport(gaze)
}
\arguments{
\item{gaze}{An object of class "ELascii", as created by readELascii().}
}
\value{
A data.frame containing a fixation report.
}
\description{
Make a fixation report from an ELascii object.
}
\details{
No details yet.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

