% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortifySRias.R
\name{fortifySRias}
\alias{fortifySRias}
\title{Fortify SRR interest areas.}
\usage{
fortifySRias(df, y1 = y1, y2 = y2)
}
\arguments{
\item{df}{A data.frame containing a set of rectangular interest
areas. Typically these will contain at least 7 columns
for: type, index x_left, y_top, x_right, y_bottom, label.}

\item{y1}{Label of column in .data containing y values for upper
bound of interest areas.}

\item{y2}{Label of column in .data containing y values for lower
bound of interest areas.}
}
\value{
A data.frame containing interest areas, now enriched with
    columns for:

\describe{
    \item{iaHeight}{Integer. Interest area height in pixels.}
    \item{firstWordInLine}{Logical. Flag the first word (IA) on each line.}
    \item{lastWordInLine}{Logical. flag the last word (IA) on each line.}
    \item{lineNum}{Integer. Line number, ranging from 1 to n where n is the number of lines.}
}
}
\description{
Fortify SRR interest areas for text stimuli with
    interest area heights, first word locations, and line numbers.
}
\details{
This function takes a data.frame containing details of
    SRR compliant rectangular interest areas and fortifies them
    with interest area heights, first word locations, and line
    numbers. The assumption is that these interest areas are
    defined over for a text stimulus (e.g., each interest area
    contains a word from a multi-line text).
}
\examples{
\dontrun{

library(dplyr)
library(readr)

ias0 <- read_tsv(system.file()) ## add an actual SRR IA file to /inst/extdata
glimpse(ias0)

ias1 <- fortifySRias(ias0, y1=y_top, y2=y_bottom)
glimpse(ias1)

}
}
\seealso{
patchIAy
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
