% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_y.R
\name{get_start_pts}
\alias{get_start_pts}
\title{Get y positions of text lines from region file}
\usage{
get_start_pts(region_file, trial_num)
}
\arguments{
\item{region_file}{A "*.csv" character list recording the name of the region file}

\item{trial_num}{An integer indicating the trial number of the current start_pts}
}
\value{
A data.frame recording the left position of the first word in each line of the text in each trial. It has three columns:

\enumerate{
    \item x_pos: x position of the first word in each line of the text.
    \item y_pos: y position of the first word in each line of the text.
    \item trial_num: Integer indicating the trial number of the current start_pts.
}
}
\description{
This function extracts (x_pos, y_pos) of first word in each text line from the region file
}
\details{
This function reads region_file and gets (x_pos, y_pos) of first word in each text line.
         Then, it bounds (x_pos, y_pos) with trail_num into a data frame, and return the data frame.
}
\author{
Tao Gong \email{gtojty@gmail.com}
}
