% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_y.R
\name{get_xybounds}
\alias{get_xybounds}
\title{Get xy_bounds from region file}
\usage{
get_xybounds(region_file, trial_num, image_width = 1280,
  image_height = 1024)
}
\arguments{
\item{region_file}{region file name (csv file)}

\item{trial_num}{the trial number of the current start_pts}

\item{image_width}{default x_max boundary (1280)}

\item{image_height}{default y_max boundary (1024)}
}
\value{
A matrix recording the eye-movement recording boundary in the trial trial_num. Each row has five columns:
 \enumerate{
    \item x_min: Minimum x value across all text lines.
    \item x_max: Maximum x value across all text lines.
    \item y_min: Minimum y value across all text lines.
    \item y_max: Maximum y value across all text lines.
    \item trial_num: Integer indicating the trial number of the current start_pts.
}
}
\description{
This function extracts (x_min, x_max, y_min, y_max) from the region file
}
\details{
This function reads region_file and gets (x_min, x_max, y_min, y_max) of all text lines.
         Then, it bounds (x_min, x_max, y_min, y_max) with trail_num into a list, and return it.
}
\author{
Tao Gong \email{gtojty@gmail.com}
}
