% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textCanvas.R
\name{getChars}
\alias{getChars}
\title{Sweep each line of text in a text cavas to estimate
    character locations.}
\usage{
getChars(canvas, lines)
}
\arguments{
\item{canvas}{A N dimensional matrix (product of png::readPNG())
corresponding to bitmap image.}

\item{lines}{Data.frame generated by applying guessLines() to
canvas.}
}
\value{
A data frame containing 1 row for each character in each
    line of canvas, and the following columns:

\enumerate{
    \item line: Integer indicating line number, counting from the top of the canvas.
    \item char: Integer indicating the character number, counting from the left most.
    \item left: Integer indicating left-most extent of character in pixels.
    \item right: Integer indicating right-most extent of character in pixels.
}
}
\author{
Dave Braze <davebraze@gmail.com>
}

