% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readELascii.R
\name{getEyelinkTrialData}
\alias{getEyelinkTrialData}
\title{Called by readELascii()}
\usage{
getEyelinkTrialData(bounds, lines)
}
\arguments{
\item{bounds}{A numeric tuple. e1 is index marking beginning of
trial. e2 is index indicating end of trial.}

\item{lines}{A vector of strings, each corresponding to 1 line of
the EL ASCII file.}
}
\value{
A list of 6 elements, data.frames enumerating fixations,
    saccades, blinks, TRIAL_VARs, samples and messages for the
    trial.
}
\description{
Used by readELascii(). Not intended for end-users.
}
\details{
Used by readELascii(). Not intended for end-users. Extract fixations, saccades, blinks
    and other data from an individual trial.
}
\author{
Dave Braze \email{davebraze@gmail.com}

Monica Li \email{monica.yc.li@gmail.com}
}
