% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textCanvas.R
\name{getMargins}
\alias{getMargins}
\title{Find the margins in a text canvas}
\usage{
getMargins(canvas)
}
\arguments{
\item{canvas}{A N dimensional matrix (product of png::readPNG())
corresponding to bitmap image.}
}
\value{
A named 4 vector containing top, right, bottom and left margins
}
\description{
None yet
}
\details{
None yet
}
\author{
Dave Braze <davebraze@gmail.com>
}

