% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegion.R
\name{getRegion}
\alias{getRegion}
\title{Find the region containing each fixation.}
\usage{
getRegion(fixReport, regionList, noRegnum = NA, noReglabel = "",
  supplement = TRUE)
}
\arguments{
\item{fixReport}{A data.frame containin a fixation Report as may
be created by fixReport(). Must contain columns \code{xpos}
and \code{ypos}.}

\item{regionList}{A data.frame containing regions of
interest. Must have columns with region bounds labeled 'left',
'right', 'top', 'bottom'. Regions are assumed to be
rectangular.}

\item{noReglabel}{The label assigned to fixations outside of any
region defined in \code{regions} argument.}

\item{supplement}{If TRUE, getRegion() will return a data.frame
    consisting of fixReport suplemented with columns regnum and
    reglabel. If FALSE, return a 2 column data.frame with just
    those two columns.

##' @return A data.frame including columns regnum and reglabel.}
}
\description{
For each fixation in a fixation report, this function
    will determine the Region of Interest that contains it, if any.
}
\details{
For each fixation in a fixation report, this function
    will find the Region of interest that contains it, if any.

    Regions are assumed to be rectangular. Region bounds labeled as
     'left', 'right', 'top', 'bottom'.

    The value of \code{supplement} determines whether to return
    the region values within an enriched version of \code{fixReport},
    or as a simple two column data.frame.
}
\author{
Dave Braze <davebraze@gmail.com>
}
