% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textCanvas.R
\name{inked}
\alias{inked}
\title{Counts 'inked' values in a vector.}
\usage{
inked(v, bg, tol = 1/(2^32))
}
\arguments{
\item{v}{A vector of pixels. It will generally be either a row or column from a text canvas.}

\item{bg}{The background value.}

\item{tol}{A tolerance around the background value. Defaults to 1/(2^32).}
}
\value{
A numeric value corresponding to the number of non-background values in \code{v}.
}
\description{
Counts number of values in a vector that are different from the 'background'.
}
\details{
Counts number of values in a vector that are different from the 'background'. Intended
    to be apply()ed to the rows or columns of a matrix (a color plane in a text canvas).

    Not necessarily intended to be called by the end user.
}
\author{
Dave Braze <davebraze@gmail.com>
}
