% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeReport.R
\name{makeReport}
\alias{makeReport}
\title{Generate an ET report, as data.frame.}
\usage{
makeReport(gaze, type = c("FIX", "SACC", "TRIALVAR"))
}
\arguments{
\item{gaze}{An object of class "ELascii".}

\item{type}{Which type of base report to generate. Ranges over c("FIX", "SACC", "TRIALVAR").}
}
\value{
A data.frame containing the requested report.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

