% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_y.R
\name{mark_oob}
\alias{mark_oob}
\title{Mark out of boundary fixations}
\usage{
mark_oob(data, xy_bounds, trial_num, cur_trial)
}
\arguments{
\item{data}{fixation data}

\item{xy_bounds}{boundary of fixations. If it has one row, all trials use the same boundary (xmin, xmax, ymin, ymax);
if it has many rows, the number of row should be equal to the number of trials.}

\item{trial_num}{number of trials}

\item{cur_trial}{the current trail index}
}
\value{
A copy of data with out of boundary fixations marked as 'oob'
}
\description{
This function is modified from Cohen's paper
}
\details{
<<<<<<< HEAD
This function marks fixations outside boundary as 'oob'
=======
This function marks fixations oustide boundary as 'oob'
>>>>>>> 2385479... adjust_y.R, edits to roxygen blocks
}
\author{
Tao Gong <gtojty@gmail.com>
}

