% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestPOI.R
\name{nearestPOI}
\alias{nearestPOI}
\title{Find the nearest Point of Interest for each fixation.}
\usage{
nearestPOI(fixReport, poiList, supplement = TRUE)
}
\arguments{
\item{fixReport}{A fixation Report as may be created by
fixReport(). Must contain columns \code{xpos} and \code{ypos}.}

\item{poiList}{A data.frame containing x,y coordinates for Points
of Interest. The data.frame must contain exactly two columns,
named \code{xpos} and \code{ypos}, in that order.}

\item{supplement}{If TRUE, nearestPOI() will return a data.frame
consisting of fixReport suplemented with columns nearestpoi
and dist2poi. If FALSE return a 2 column data.frame with just
those two columns.}
}
\value{
A data.frame including columns nearestpoi and dist2poi.
}
\description{
For each fixation in a fixation report, this function
    will determine the nearest Point of Interest and the
    fixation's distance from that Point of Interest.
}
\details{
For each fixation in a fixation report, this function
    will determine the nearest Point of Interest and the
    fixation's distance from that Point of Interest.

    The value of \code{supplement} determines whether to return
    the POI values within an enriched version of \code{fixReport},
    or as a simple two column data.frame.
}
\author{
Dave Braze <davebraze@gmail.com>
}
