% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchIAy.R
\name{patchIAy}
\alias{patchIAy}
\title{Modify upper and lower limits of Interest Areas.}
\usage{
patchIAy(df, y1 = y1, y2 = y2, lineNum = lineNum, stdize = TRUE,
  ytop = 0, ybottom = 0)
}
\arguments{
\item{df}{A data.frame containing a set of rectangular interest
areas conforming to SRR standards. It will contain 7 columns
for: type, index x_left, y_top, x_right, y_bottom, label. Note
that SRR *ias files are TSV files without column
labels. Labels should be added when reading the file.}

\item{y1}{Label of column in df containing y values for upper
bound of interest areas.}

\item{y2}{Label of column in df containing y values for lower
bound of interest areas.}

\item{lineNum}{Label of column in df indicating line number for
each Interest Area.}

\item{stdize}{Logical. Indicates whether or not to do line-wise standardization
of IA upper and lower limits. Defaults to TRUE.}

\item{ytop}{Integer. Offset to apply to upper limit of IAs. Positive values increase
the height of IAs (moves upper limit up). Defaults to 0.}

\item{ybottom}{Integer. Offset to apply to lower limit of IAs. Positive values increase
the height of IAs (moves lower limit down). Defaults to 0.}
}
\value{
A data.frame of interest areas with modified upper and
    lower limits.
}
\description{
Modify upper and lower limits of Interest Areas.
}
\details{
Modify upper and lower limits of Interest
    Areas. Modification is done in place in the sense that the
    returned data.frame has the same columns (and same number of
    columns) as the one passed to the function, but previously
    existing upper and lower bounds for interest areas are
    over-written with new values.
}
\examples{
\dontrun{

library(dplyr)
library(readr)

ias0 <- read_tsv(system.file()) ## add an actual SRR IA file to /inst/extdata
glimpse(ias0)

ias1 <- fortifySRias(ias0, y1=y_top, y2=y_bottom)
glimpse(ias1)

ias2 <- patchIAy(ias1)
glimpse(ias2)

}
}
\seealso{
fortifySRias
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
