% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.ELascii.R
\name{read.ELascii}
\alias{read.ELascii}
\title{Get events from SR Research ASCII data files.}
\usage{
read.ELascii(file, tStartRE = "TRIALID", tEndRE = "TRIAL_RESULT",
  eye = NA)
}
\arguments{
\item{file}{A string giving path/fname to input file (ELalscii file).}

\item{tStartRE}{A string containing regular expression that uniquely identifies beginning of trial.}

\item{tEndRE}{A string containing regular expression that uniquely identifies end of trial.}

\item{eye}{Indicates which eye ("R"|"L") to get events from. Currently unused.}
}
\value{
List with one element for the file header and one element for each trial. Each trial
element is itself a list of 4 elements: data.frames enumerating fixations, saccades, blinks and
TRIAL_VARs for the trial.
}
\description{
SR Research provides a utility (EDF2ASC.exe) that dumps ASCII renderings of their proprietary
EDF data file format. This function reads those ASCII files and extracts eye-movement events
(fixations, saccades, blinks) and TRIAL_VARs from them.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

