% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textRegions.R
\name{reg2regdef}
\alias{reg2regdef}
\title{Convert region file to region definition file.}
\usage{
reg2regdef(reg, scrnW = NA, scrnH = NA, fnt.name = NA, fnt.size = NA,
  chrW = NA, chrH = NA, ln.space = NA, baseline = NA, mrgn.top = NA,
  mrgn.left = NA, mrgn.bottom = NA, mrgn.right = NA, rgn.maxH = NA,
  rgn.minH = NA, rgn.padL = NA, rgn.padR = NA)
}
\arguments{
\item{reg}{A data.frame containing region specifications, as read
from a region file ("*.region.csv").}

\item{scrnW}{Screen width in pixels (integer).}

\item{scrnH}{Screen height in pixels (integer).}

\item{fnt.name}{Font name used for stimulus text.}

\item{fnt.size}{Nominal font size in points for text display.}

\item{chrW}{Letter width in pixels.}

\item{chrH}{Nominal letter height in pixels.}

\item{ln.space}{Line spacing in pixels for multi line texts. Multi
line texts are not currently supported.}

\item{baseline}{Baseline positions for each line of text. Measured
in pixels from the top of the screen. Multi line texts are not
currently supported.}

\item{mrgn.top}{Top margin in pixels.}

\item{mrgn.left}{Left margin in pixels.}

\item{mrgn.bottom}{Bottom margin in pixels.}

\item{mrgn.right}{Right margin in pixels.}

\item{rgn.maxH}{Extent of regions of interest above baseline in
pixels.}

\item{rgn.minH}{Extent of regions of interest below baseline in
pixels.}

\item{rgn.padL}{Expand leftmost region on each line leftward by
this amount in pixels.}

\item{rgn.padR}{Expand rightmost region on each line rightward by
this amount in pixels.}
}
\value{
A vector of strings containing the region definition. The
    vector includes a yaml block with values for each of the
    function parameters except for "reg". In addition to the yaml
    block, the vector will include a pair of lines for each line
    of text in the stimulus. The first element of each pair is the
    text displayed on that line. The second element is a regioning
    string made up of dots ("."), and pipe ("|")
    characters. Pipes indicate the beginnings of regions. By
    default, the region definition file will specify that each
    text line be exhaustively dividied into space delimited regions
    (i.e. there will be a pipe character corresponding to each space
    character in the paired text line.

    This vector can be written to file and hand edited to add or
    correct information in the yaml block, or to re-specify region
    placements.
}
\description{
Convert a file containing full region description
    into a region definition file. The latter is suitable for hand
    editing and can be used to generate alternative region
    specifications (e.g., multi word regions) for text stimuli.
}
\details{
We aspire to handle input files (region files) for multi
    line texts, but at present only region files for single line
    texts are handled.

    File parameter values are used to fill in parameters written
    to the yaml block of the region definition file. Note that few
    of these parameters are important when the region definition
    file is used to create a new region file. Two parameters are
    critical to proper region definitions.

    The chrW parameter is used to translate region boundaries in x
    dimension from letter positions (as specified in the region
    definition file) to pixel positions (as required for the
    region or ias file). chrW is estimated from contents of
    region.csv file, and will probably be correct most of the
    time. Regardless, it should be checked and, if necessary,
    manually edited in the resulting region definition file.

    Accurate baseline positions are also critical to determining
    the y positionins of regions. Baselines are read directly from
    the region.csv file and should be accurate. Note that baseline
    positions, in pixels, are measured from the TOP of the screen.

    ToDo: Make to work with region files for multi-line text
    stims.

    ToDo: Make to read/parse SRR IAS files and build region defs
    based on them.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

