% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textRegions.R
\name{regdef2ias}
\alias{regdef2ias}
\title{Convert region definition file to SRR Interest Area file.}
\usage{
regdef2ias(fname)
}
\arguments{
\item{fname}{A string containing the name of a "region definition"
file, such as might be created by reg2regdef(). See Details.}
}
\value{
A data.frame specifying the content of an SRR interest
    area file (*.ias).  Use \code{readr::write_delim(...,
    delim="\n", col_names=FALSE)} to save the interest area
    specification to file.
}
\description{
Convert a region definition file to an SRR Interest
    Area file (*.ias). The latter can be hand edited to specify
    alternative region specifications (e.g., multi word regions)
    for text stimuli.
}
\details{
We aspire to handle region definitions for multi line
    texts, but at present only region files for single line texts
    are handled.

    Parameter values are read from the yaml block of the region
    definition file. Note that a few of these parameters are
    important when translating region definitions to interest
    areas.

    A region definition file contains 2 parts. The first is a yaml
    block with values for each of the function parameters except
    for "reg". Three parameters in the yaml block are critical to
    proper region definitions.

    The chrW parameter is used to translate region boundaries in x
    dimension from letter positions (as specified in the region
    definition file) to pixel positions (as required for the
    region or ias file). chrW is estimated from contents of
    region.csv file, and will probably be correct most of the
    time. Regardless, it should be checked and, if necessary,
    manually edited in the resulting region definition file.

    Accurate baseline positions are also critical to determining
    the y positionins of regions. Baselines are read directly from
    the region.csv file and should be accurate. Note that baseline
    positions, in pixels, are measured from the TOP of the screen.

    Finally, left margin specification is important as this gives
    an x offset for region positions.

    Four additional parameters in the yaml block of the region
    definition file will be used to modify regions. They are:
    regions$maxH, regions$minH, regions$padL, and regions$padR.

    The second part of a region definition file is the region
    block. This block contains a pair of lines for each line of
    text in the stimulus. The first element of each pair is the
    text displayed on that line. The second element is a regioning
    string made up of dots ("."), and pipe ("|") characters. Pipes
    indicate the beginnings of regions. By default, the region
    definition file will specify that each text line be
    exhaustively dividied into space delimited regions (i.e. there
    will be a pipe character corrponding to each space character
    in the paired text line.

    The region definition file can be hand edited to add or
    correct information in the yaml block, or to re-specify region
    placements.

    ToDo: Make to work with region definition files for multi-line
    text stims.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
\seealso{
reg2regdef
}

