% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region2SRias.R
\name{region2SRias}
\alias{region2SRias}
\title{Convert region file to SR interest area file}
\usage{
region2SRias(regfile, iasfile = NULL, size = c("big", "small"),
  yoffsets = NULL, xpad = NULL)
}
\arguments{
\item{regfile}{Region file to read, formatted per Tao Gong's
Python code.}

\item{iasfile}{If NULL (default) then create the output filename
by stripping ext from regfile and adding ".ias". If not NULL,
then this should be a string specifying output filename.}

\item{size}{Extract "big" or "small" interest areas from region
file.}

\item{yoffsets}{If NULL (default), use y coordinates as read from
region file for Interest Areas. If not NULL, must be a
2-vector of integers indicating extent above and below
baseline for Interest areas (in pixels). Both values should be
positive. If yoffsets is not NULL, then the 'size' argument is
ignored.}

\item{xpad}{If NULL (default), use x coordinates for Interest
Areas as read from region file. If not NULL, must be a
2-vector of integers indicating extent leftward and rightward
to expand the Interest Areas for the first and last words on
each line (in pixels). Both values should be positive.}
}
\value{
Invisibly returns a data.frame containing a set of
    Interest Areas, but typically called for its side effect of
    writing an "*.ias" file.
}
\description{
Convert a region file to an SR interest area file. We
    only handle static IAS files.
}
\details{
No details yet.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

