% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeReport.R
\name{reportTrials}
\alias{reportTrials}
\title{Extract a trial report from an ELascii object}
\usage{
reportTrials(dat)
}
\arguments{
\item{dat}{An object of class "ELascii", as created by readELascii().}
}
\value{
Which specific trial variables are available will vary, depending on the idiosyncratic
    characteristics of the protocol. Regardless, the returned data.frame will contain a report with the
    following columns, together with any trial variables found in the dataset:

\enumerate{
    \item subject: factor for subject ID, taken from the name of the source EDF file.
    \item trialn: factor for trial ID.
    \item time0: integer indicating timestamp (ms) of first sample in recording block.
}

Returned data.frame will also include trial_vars from edf file if requested (TRUE by default).
}
\description{
Extract a trial report from an ELascii object
}
\details{
trial reports contain summary data from each trial, as well as subject ID,
    trial number.
}
\examples{
fname <- system.file("/extdata/1950006-RAN.asc.gz", package="FDBeye")
e <- readELascii(fname)
efix <- reportTrials(e)
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
