% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeReport.R
\name{saccReport}
\alias{saccReport}
\title{Extract a saccade report from an ELascii object}
\usage{
saccReport(dat, trialvars = TRUE)
}
\arguments{
\item{dat}{An object of class "ELascii", as created by readELascii().}

\item{trialvars}{Logical, defaults TRUE to include trial variables in saccade report.}
}
\value{
A data.frame containing a saccade report with the following columns:

\enumerate{
    \item subject: factor for subject ID, taken from the name of the source EDF file.
    \item trialn: factor for trial ID.
    \item time0: integer indicating timestamp (ms) of first sample in recording block.
    \item event: factor for event type, always "SACC".
    \item eye: factor indicating eye for corresponding saccade, ranges over c("L", "R").
    \item stime: integer indicating start time for fixation (ms).
    \item etime: integer indicating end time for fixation (ms).
    \item dur: integer indicating duration of fixation (ms).
    \item xpos1: starting x position
    \item ypos1: starting y position
    \item xpos2: ending x position
    \item ypos2: ending y position
    \item ampl: amplitude (units?)
    \item peakvel: peak velocity (units?)
}

Returned data.frame will also include trial_vars from edf file if requested (TRUE by default).
}
\description{
Extract a saccade report from an ELascii object
}
\details{
Saccade reports include data from both eyes, if available, as well as subject ID,
    trial number, time stamp of first sample in each trial. By default, values for trial
    variables found in each trial are also include, although these may be suppressed.
}
\examples{
fname <- system.file("/extdata/1950006-RAN.asc.gz", package="FDBeye")
e <- readELascii(fname)
esacc <- saccReport(e)
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

