% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_y.R
\name{trial_plots}
\alias{trial_plots}
\title{This function draws original and modified fixations with fitted lines}
\usage{
trial_plots(data, start_pts, output_filehead, draw_type, bg_image_name = NULL,
  image_width = 1280, image_height = 1024)
}
\arguments{
\item{data}{fixation data frame containing positions, durations, fitted measure and fitted line information}

\item{start_pts}{starting point of each base line (x_pos, y_pos)}

\item{output_filehead}{output figure file head}

\item{draw_type}{'hand' -> hand-made fixation lines, output file ends with '_hand'
'original' -> original fixation lines, output file ends with '_ori'
'modified' -> residule fixation lines, output file ends with '_mod'}

\item{bg_image_name}{background trial figure file}

\item{image_width}{default image width (1280)}

\item{image_height}{default image height (1024)}
}
\value{
None. Called for the side effect of creating a plot.
}
\description{
This function is suitable for values optimized by create_lines3 and create_lines4
}
\details{
TAO, PLEASE PROVIDE DETAILS
}
\author{
Tao Gong <gtojty@gmail.com>
}

