% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edf2asc.R
\name{edf2asc}
\alias{edf2asc}
\title{Call "edf2asc" command line utility to do some work.}
\usage{
edf2asc(edffiles, e2afname = "edf2asc64.exe",
  e2apath = "D:/winbin/SR Research/EyeLink/EDF_Access_API/Example")
}
\arguments{
\item{edffiles}{Vector of *edf file names.}

\item{e2afname}{Name of edf2asc executable (e.g., "edf2asc.exe", "edf2asc64.exe").}

\item{e2apath}{Path to edf2asc executable. It will probably look something like
"D:/winbin/SR Research/EyeLink/EDF_Access_API/Example" from "SR Research" on down.}

\item{files}{A character vector listing file names to be converted.}
}
\value{
A character vector listing output files.
}
\details{
Call SR Research "edf2asc" command line utility to convert files.

\enumerate{
    \item The best way to get the edf2asc utility is to install the Eyelink Developers Kit: https://www.sr-support.com/showthread.php?6-EyeLink-Developers-Kit-for-Windows-%28Windows-Display-Software%29.
    \item Documentation is in the EL1000+ manual, section 4.8 "Using ASC files".
    \item Make sure edfapi library (e.g., edfapi.dll) is somewhere on the PATH.
}
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

