% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixPlot.R
\name{fixPlot}
\alias{fixPlot}
\title{Plot Fixations over bitmap using ggplot2.}
\usage{
fixPlot(bgimage, bgalpha = 0.33, mar = c(0, 0, 0, 0), data)
}
\arguments{
\item{bgimage}{Path to background image for the plot. Usually will be the visual stimulus
subjects viewed will eye movements were recorded.}

\item{bgalpha}{Alpha level for background layer. Currently not working.}

\item{mar}{A 4 vector for margin adjustment: top, right, bottom, left. Use positive values move
margins toward center of plot (trim the display area). Use negative values to expand the
display area beyond the bitmap.}

\item{data}{A data.frame containing fixations. Must include x and y positions with column names
"x" and "y".}
}
\value{
A ggplot2 object.
}
\description{
A convenience wrapper around ggplot2::ggplot()
}
\details{
This is a convenience function for using ggplot() to lay fixation data over a
    background bitmap.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

