% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.ELascii.R
\name{getEyelinkTrialData}
\alias{getEyelinkTrialData}
\title{Used by read.ELascii(). Not intended for end-users.}
\usage{
getEyelinkTrialData(bounds, lines)
}
\arguments{
\item{bounds}{A numeric tuple. e1 is index marking beginning of trial. e2 is index indicating
end of trial.}

\item{lines}{A vector of strings, each corresponding to 1 line of the EL ASCII file.}
}
\value{
A list of 4 elements, data.frames enumerating fixations, saccades, blinks and TRIAL_VARs for the
trial.
}
\description{
Used by read.ELascii(). Not intended for end-users. Extract fixations, saccades, and blinks from a trial.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

