% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textCanvas.R
\name{getMargins}
\alias{getMargins}
\title{Find the margins in a text canvas}
\usage{
getMargins(canvas)
}
\arguments{
\item{canvas}{A N dimensional matrix (product of png::readPNG())
corresponding to bitmap image.}
}
\value{
A named 4 vector containing top, right, bottom and left margins
}
\description{
None yet
}
\details{
None yet
}
\examples{
    cnvs <- system.file("extdata/story01.png", package="FDBeye")
    cnvs <- png::readPNG(cnvs)
    fcnvs <- apply(cnvs, c(1,2), sum) # flatten to a single plane for convenience

    ## get margins
    getMargins(fcnvs)
}
\author{
Dave Braze <davebraze@gmail.com>
}

