% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_y.R
\name{mark_oob}
\alias{mark_oob}
\title{Mark out of boundary fixations}
\usage{
mark_oob(data, xy_bounds, trial_num, cur_trial)
}
\arguments{
\item{data}{A data.frame containing the fixation data}

\item{xy_bounds}{A mtraix containing the boundaries of fixations.
If it has only one row, all trials use the same boundary.
If it has many rows, the number of row should be equal to the number of trials,
and each row corresponds to the boundary of the corresponding trial.
In each row, it has four columns:
\enumerate{
  \item x_min: Left boundary (minimum x value within the boundary).
  \item x_max: Right boundary (maximum x value within the boundary).
  \item y_min: Top boundary (minimum y value within the boundary).
  \item y_max: Bottom boundary (maximum y value within the boundary).
}}

\item{trial_num}{Integer indicating the total number of trials}

\item{cur_trial}{Integer indicating the current trail}
}
\value{
A copy of data with out of boundary fixations marked as 'oob' in column type
}
\description{
This function is modified from Cohen's paper
}
\details{
This function marks fixations outside boundary as 'oob'
}
\author{
Tao Gong \email{gtojty@gmail.com}
}

