% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textRegions.R
\name{regdef2ias}
\alias{regdef2ias}
\title{Convert region file to region definition file.}
\usage{
regdef2ias(fname)
}
\arguments{
\item{fname}{A string containing the name of a "region definition"
file, such as might be created by reg2regdef(). See Details.}
}
\value{
A vector of strings containing the region definition. The
    vector includes a yaml block with values for each of the
    function parameters except for "reg". In addition to the yaml
    block, the vector will include a pair of lines for each line
    of text in the stimulus. The first element of each pair is the
    text displayed on that line. The second element is a regioning
    string made up of spaces (" "), and pipe ("|")
    characters. Pipes indicate the beginnings of regions. By
    default, the region definition file will specify that each
    line be exhaustively dividied into space delimited regions
    (i.e. there will be a pipe character corrponding to each space
    character in the paired text line.

    This vector can be written to file and hand edited to add or correct information in the yaml
    block, or to re-specify region placements.
}
\description{
Convert a file containing full region description
    into a region definition file. The latter is suitable for hand
    editing and can be used to generate alternative region
    specifications (e.g., multi word regions) for text stimuli.
}
\details{
We aspire to handle region definitions for multi line
    texts, but at present only region files for single line texts
    are handled.

    File parameter values are used to fill in parameters written
    to the yaml block of the region definition file. Note that few
    of these parameters are important when the region definition
    file is used to create a new region file.

    A region definition file contains 2 parts. The first is a yaml
    block with values for each of the function parameters except
    for "reg".  Two parameters in the yaml block are critical to
    proper region definitions.  The chrW parameter is used to
    translate region boundaries in x dimension from letter
    positions (as specified in the region definition file) to
    pixel positions (as required for the region or ias file). chrW
    is estimated from contents of region.csv file, and will
    probably be correct most of the time. Regardless, it should be
    checked and, if necessary, manually edited in the resulting
    region definition file.

    Accurate baseline positions are also critical to determining
    the y positionins of regions. Baselines are read directly from
    the region.csv file and should be accurate. Note that baseline
    positions, in pixels, are measured from the TOP of the screen.

    The second part of a region definition file is the region
    block. This block contains a pair of lines for each line of
    text in the stimulus. The first element of each pair is the
    text displayed on that line. The second element is a regioning
    string made up of spaces (" "), and pipe ("|")
    characters. Pipes indicate the beginnings of regions. By
    default, the region definition file will specify that each
    line be exhaustively dividied into space delimited regions
    (i.e. there will be a pipe character corrponding to each space
    character in the paired text line.

    The region definition file can be hand edited to add or
    correct information in the yaml block, or to re-specify region
    placements.

    ToDo: Make to work with region definition files for multi-line
    text stims.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

