% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region2SRias.R
\name{region2SRias}
\alias{region2SRias}
\title{Convert region file to SR interest area file.}
\usage{
region2SRias(regfile, iasfile = FALSE, size = c("big", "small"))
}
\arguments{
\item{regfile}{Region file to read, formatted per Tao Gong's Python code.}

\item{iasfile}{If not FALSE (default), then this is a string specifying output filename. If
FALSE then create the output filename by stripping ext from regfile and adding ".ias".}

\item{size}{Extract "big" or "small" interast areas from region file.}
}
\value{
Called for its side effect of writing an "*.ias" file.
}
\description{
Convert a region file to an SR interest area file.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

